[WFCORE-3119] Move validators into rescource definition and reduce visibility.
