--- elytron/src/main/java/org/wildfly/extension/elytron/validators/SizeValidator.java
@@ -1,85 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2017 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wildfly.extension.elytron.validators;
-
-import org.jboss.as.controller.OperationFailedException;
-import org.jboss.as.controller.operations.validation.ModelTypeValidator;
-import org.jboss.dmr.ModelNode;
-import org.jboss.dmr.ModelType;
-import org.wildfly.extension.elytron._private.ElytronSubsystemMessages;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-/**
- * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
- */
-public class SizeValidator extends ModelTypeValidator {
-    private static final Pattern SIZE_PATTERN = Pattern.compile("(\\d+)([kKmMgGbBtT])?");
-
-    public SizeValidator() {
-        this(false);
-    }
-
-    public SizeValidator(final boolean nullable) {
-        super(ModelType.STRING, nullable);
-    }
-
-    @Override
-    public void validateParameter(String parameterName, ModelNode value) throws OperationFailedException {
-        super.validateParameter(parameterName, value);
-        if (value.isDefined()) {
-            parseSize(value);
-        }
-    }
-
-    public static long parseSize(final ModelNode value) throws OperationFailedException {
-        final Matcher matcher = SIZE_PATTERN.matcher(value.asString());
-        if (!matcher.matches()) {
-            throw ElytronSubsystemMessages.ROOT_LOGGER.invalidSize(value.asString());
-        }
-        long qty = Long.parseLong(matcher.group(1), 10);
-        final String chr = matcher.group(2);
-        if (chr != null) {
-            switch (chr.charAt(0)) {
-                case 'b':
-                case 'B':
-                    break;
-                case 'k':
-                case 'K':
-                    qty <<= 10L;
-                    break;
-                case 'm':
-                case 'M':
-                    qty <<= 20L;
-                    break;
-                case 'g':
-                case 'G':
-                    qty <<= 30L;
-                    break;
-                case 't':
-                case 'T':
-                    qty <<= 40L;
-                    break;
-                default:
-                    throw ElytronSubsystemMessages.ROOT_LOGGER.invalidSize(value.asString());
-            }
-        }
-        return qty;
-    }
-}
